backend/requirements.txt
fastapi
uvicorn
sqlalchemy
playwright
pandas
openpyxl
python-multipart
reportlab
requests
beautifulsoup4
gspread
google-auth
google-cloud-storage
python-dotenv
ddgs
psycopg2-binary
